requirements.txt
google_api_python_client==2.170.0
google_auth_oauthlib==1.2.1
joblib==1.5.1
numpy==2.3.1
pandas==2.3.0
protobuf==6.31.1
PyMySQL==1.1.1
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
pytz==2024.2
slack_sdk==3.35.0
SQLAlchemy==2.0.41
